[difftest] refactor spike events
